requirements.txt
asgiref==3.8.1
Django==4.2.18
et-xmlfile==2.0.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2024.2
